feat: finish the clinicalRecord with all steps done
