+++ src/config/routes.jsx
@@ -28,7 +28,6 @@ import StudiesRequest from "../pages/StudiesRequest/StudiesRequest";
 import Reports from "../pages/VisitReports/VisitReports";
 import PetsReports from "../pages/PetsReports/PetsReports";
 
-import ClinicalRecordT from "../pages/ClinicalRecordT/Clinical RecordT";
 
 //Client pages
 import Landing from "../pages/Landing/Landing";
@@ -107,13 +106,8 @@ const routesAdmin = [
         layout: LayoutAdmin,
     },
     {
-        path: "/clinical-record",
+        path:"/clinical-record",
         component: ClinicalRecord,
-        layout: LayoutAdmin,
-    },
-    {
-        path:"/clinical-recordT",
-        component: ClinicalRecordT,
         layout:LayoutAdmin,
     },
     {
