+++ src/pages/ClinicalRecord/ClinicalRecord.scss
@@ -1,49 +1,60 @@
 @import "../../scss/index.scss";
 
-.motive{
-    margin-top: 1%!important;
-    color: $primary-color-hover!important;
+.motive {
+  margin-top: 1% !important;
+  color: $primary-color-hover !important;
 }
-.example-link{
-    margin-right: 16px;
-    line-height: 24px;
+.example-link {
+  margin-right: 16px;
+  line-height: 24px;
+}
 
-}
-// .steps{
-//     // margin-top: 2%;
-// }
 .steps-content {
-    min-height: 200px;
-    margin-top: 16px;
-    padding-top: 30px;
-    text-align: center;
-    background-color: #fafafa;
-    border: 1px dashed #e9e9e9;
-    border-radius: 2px;
-  }
-  
-  .steps-action {
-    margin-top: 24px;
-    display: flex;
-    justify-content: center;
+  min-height: 200px;
+  margin-top: 16px;
+  padding-top: 30px;
+  text-align: center;
+  background-color: #fafafa;
+  border: 1px dashed #e9e9e9;
+  border-radius: 2px;
+}
 
-    &-next{
-        height: 60px;
-        width: 60px;
-        font-size: 25px;
-        border-radius: 30px;
-    }
-    &-back{
-        height: 60px;
-        width: 60px;
-        font-size: 25px;
-        border-radius: 30px;
-    }
-    &-finish{
-        height: 60px;
-        width: 60px;
-        font-size: 25px;
-        border-radius: 30px;
-    }
+.steps-action {
+  margin-top: 24px;
+  display: flex;
+  justify-content: center;
 
+  &-next {
+    height: 60px;
+    width: 60px;
+    font-size: 25px;
+    border-radius: 30px;
   }
+  &-back {
+    height: 60px;
+    width: 60px;
+    font-size: 25px;
+    border-radius: 30px;
+  }
+  &-finish {
+    height: 60px;
+    width: 60px;
+    font-size: 25px;
+    border-radius: 30px;
+  }
+}
+.goBackButton {
+  border: none;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 2vw;
+  margin-top: 6px;
+  margin-left: 1%;
+//   margin-right: 4%;
+  color : $secondary-color;
+
+  &:hover {
+    color: $primary-color-hover;
+  }
+}
